notification_service/requirements.txt
asgiref==3.11.0
Django==4.2.27
djangorestframework==3.16.1
kafka-python==2.3.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
waitress==3.0.2
